Fixed Issue#578:  cookbook installer scripts do not take a parameter for datasource-boot-script Now all installation scripts use datasource-boot-script option
Fixed Issue#581: cookbook scripts don't play well with MySQL 5.6
Changes: the variable 'MYSQL' is defined in BOOTSTRAP.sh, and as part of the post_installation actions the installer creates an option file inside the cookbook directory.
This file is used by every call to MySQL inside the cookbook (except the scripts dealing with native replication).